CR - Move hooks under own folder
